Refactor mp into mp specific directory. builds
