+++ src/common/inc/fnio.h
