+++ src/common/lib/bounce/bounce.vcxproj
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
-    <ProjectGuid>{d07edeb2-006f-4b32-8cd1-2b8fabd5879e}</ProjectGuid>
-    <TargetName>fnio</TargetName>
+    <ProjectGuid>{4a8c9c36-9b11-4bcf-9c48-0f33d4470d66}</ProjectGuid>
+    <TargetName>bounce</TargetName>
     <UndockedType>drvlib</UndockedType>
     <UndockedDir>$(SolutionDir)submodules\undocked\</UndockedDir>
     <UndockedOut>$(SolutionDir)artifacts\</UndockedOut>
@@ -11,24 +11,17 @@
   <Import Project="$(UndockedDir)vs\windows.undocked.props" />
   <Import Project="$(SolutionDir)src\wnt.cpp.props" />
   <ItemGroup>
-    <ClCompile Include="enqueue.c" />
-    <ClCompile Include="filter.c" />
-    <ClCompile Include="ioctlbounce.c" />
+    <ClCompile Include="bounce.c" />
   </ItemGroup>
   <ItemDefinitionGroup>
     <ClCompile>
       <AdditionalIncludeDirectories>
-        $(SolutionDir)inc;
-        $(SolutionDir)src\inc;
-        $(SolutionDir)src\sys\inc;
-        $(SolutionDir)submodules\ndis-driver-library\src\include;
-        $(SolutionDir)submodules\net-offloads\include;
+        $(SolutionDir)src\common\inc;
         %(AdditionalIncludeDirectories)
       </AdditionalIncludeDirectories>
       <PreprocessorDefinitions>
         POOL_NX_OPTIN_AUTO=1;
         POOL_ZERO_DOWN_LEVEL_SUPPORT=1;
-        NDIS684=1;
         %(PreprocessorDefinitions)
       </PreprocessorDefinitions>
     </ClCompile>
