+++ src/mp/sys/exclusive.h
