+++ src/mp/sys/rss.c
