+++ src/mp/sys/shared.c
