+++ src/mp/sys/shared.h
